Edits on beachballs, bounces
Moved beachball 2 & 3, to be fixed to bottom of page also put their divs in front of the palms trees in order to still be able to activate their hover. Bouncing beachball1 is now redone to bounce to percentages which allows it to still bounce on screen when the screen is different sizes or adjusted.