Edit the z-index
Moved the player behind the name tabs as well as the bouncing ball. Palms also slightly adjusted z-index. Completely changed how the player was also positioned and am using a fixed positioning with a percentage now. The sun no longer requires a top margin as a trickle effect. Also moved the player div to the lower portion of the html as added a class just for the iframe itself.